+++ tests/test_app/Plugin/AnotherTestPlugin/tests/TestCase/MyExampleTest.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace AnotherTestPlugin\MyPlugin\Test\TestCase;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * An example class test
+ */
+class MyExampleTest extends TestCase
+{
+}